services/ai-orchestrator/requirements.txt
python-dotenv
fastapi[all]
uvicorn[standard]
pydantic
pydantic_settings
pika
google-generativeai
httpx